+++ CHANGELOG.md
@@ -5,7 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/USGS-WiM/stnweb2/tree/dev) - 2020-11-10
+## [Unreleased](https://github.com/USGS-WiM/stnweb2/tree/dev)
+
+### Added
+
+-   Added esri-leaflet geosearch
+-   Added measure tools
+-   Added layout to map filters
+-   Added state, sensor type, and network name services and filter selections
+-   Added supplementary layers to toggle
+-   Added legend
+-   Added Watershed layer to legend
+
+### Changed
+
+### Fixed
+
+## [v0.1.0](https://github.com/USGS-WiM/stnweb2/releases/tag/v0.1.0) - 2020-11-10
 
 ### Added
 
@@ -17,6 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Added github super-linter
 -   Added interfaces
 -   Added esri-leaflet geosearch
+-   Added measure tools
 
 ### Changed